feat: change Video size type double -> long & add ImageRoadMap + change directory structure

+++ SpoofFileInfo/FileMetadata/VideoMetadata.cs
@@ -6,6 +6,6 @@ public class VideoMetadata : IFileMetadata
     public string Extension { get; set; } = string.Empty;
     public FileType FileType { get; init; } = FileType.Image;
     public long Size { get; set; }
-    public double Width { get; set; }
-    public double Height { get; set; }
+    public long Width { get; set; }
+    public long Height { get; set; }
 }
